feat: add online support with server, migrate to monorepo
Adding online capabilities, which are currently very basic, made it necessary to have a
client-server structure for communication bewteen parties. A new express.js project has been created
using socket.io for exchanging messages. Due to the existence of this new project, the project has
been migrated to a monorepo structure to make it easier to maintain and deploy the projects together