feat: extract shared code to project monorepo
The shared code for which we want to keep a consistency between frontend and backend (currently only
validators) has been exported to a new "packages/shared" project. The `packages/server` project is
now using `type: "module"` in the `package.json` file. There are still minor things to change, such
as importing the `shared` project without having to write the entire relative path