feat: allow users to choose their color
A user can now choose a color via a color picker in the home page before joining a game. He/she can
also change his/her color inside a room by clicking on his/her name. This color will be used for the
`Bars` as a distinctive, and also in the list of users in the top right corner