feat: add distinctive color to keys when pressed
Keys now acquire the color of the user who presses them so that it is easier to see which keys are
being pressed. Previously, users could only rely on the key bounce animation. In a future there will
be a toggle to enable/disable this functionality, as well as trails